[REFACTOR] validator를 한 번에 처리해도 될 내용을 여러번 실행하는 코드 수정 -2 #39
